requirements.txt
slack_sdk
requests
websocket-client
flask
slackeventsapi
aiohttp
python-dotenv
